+++ src/env/node/portfinder.ts
@@ -1,0 +1,6 @@
+const portfinder = require('portfinder');
+
+export const getPort = async (port: number): Promise<number> => {
+	portfinder.setBasePort(port);
+	return portfinder.getPortPromise();
+};